requirements.txt
djangorestframework
markdown       
django-filter 
scikit-learn
django-cors-headers
joblib
pillow
torch
torchvision
torchaudio
numpy
transformers
easyocr